Refactoring - move Sign.sign method into ECKeyPair so it can be overridden. This will make it easier to override the default signing behavior, such as implementing an ECKeyPair that does out of process signing.
