+++ crypto/src/main/java/org/web3j/crypto/ECKeyPair.java
@@ -4,6 +4,10 @@ import java.math.BigInteger;
 import java.security.KeyPair;
 import java.util.Arrays;
 
+import org.bouncycastle.crypto.digests.SHA256Digest;
+import org.bouncycastle.crypto.params.ECPrivateKeyParameters;
+import org.bouncycastle.crypto.signers.ECDSASigner;
+import org.bouncycastle.crypto.signers.HMacDSAKCalculator;
 import org.bouncycastle.jcajce.provider.asymmetric.ec.BCECPrivateKey;
 import org.bouncycastle.jcajce.provider.asymmetric.ec.BCECPublicKey;
 
@@ -29,6 +33,20 @@ public class ECKeyPair {
         return publicKey;
     }
 
+    /**
+     * Sign a hash with the private key of this key pair.
+     * @param transactionHash   the hash to sign
+     * @return  An {@link ECDSASignature} of the hash
+     */
+    public ECDSASignature sign(byte[] transactionHash) {
+        ECDSASigner signer = new ECDSASigner(new HMacDSAKCalculator(new SHA256Digest()));
+
+        ECPrivateKeyParameters privKey = new ECPrivateKeyParameters(privateKey, Sign.CURVE);
+        signer.init(true, privKey);
+        BigInteger[] components = signer.generateSignature(transactionHash);
+
+        return new ECDSASignature(components[0], components[1]).toCanonicalised();
+    }
 
     public static ECKeyPair create(KeyPair keyPair) {
         BCECPrivateKey privateKey = (BCECPrivateKey) keyPair.getPrivate();
