+++ crypto/src/main/java/org/web3j/crypto/Sign.java
@@ -31,17 +31,16 @@ import static org.web3j.utils.Assertions.verifyPrecondition;
 public class Sign {
 
     private static final X9ECParameters CURVE_PARAMS = CustomNamedCurves.getByName("secp256k1");
-    private static final ECDomainParameters CURVE = new ECDomainParameters(
+    static final ECDomainParameters CURVE = new ECDomainParameters(
             CURVE_PARAMS.getCurve(), CURVE_PARAMS.getG(), CURVE_PARAMS.getN(), CURVE_PARAMS.getH());
-    private static final BigInteger HALF_CURVE_ORDER = CURVE_PARAMS.getN().shiftRight(1);
+    static final BigInteger HALF_CURVE_ORDER = CURVE_PARAMS.getN().shiftRight(1);
 
     public static SignatureData signMessage(byte[] message, ECKeyPair keyPair) {
-        BigInteger privateKey = keyPair.getPrivateKey();
         BigInteger publicKey = keyPair.getPublicKey();
 
         byte[] messageHash = Hash.sha3(message);
 
-        ECDSASignature sig = sign(messageHash, privateKey);
+        ECDSASignature sig = keyPair.sign(messageHash);
         // Now we have to work backwards to figure out the recId needed to recover the signature.
         int recId = -1;
         for (int i = 0; i < 4; i++) {
@@ -64,16 +63,6 @@ public class Sign {
         byte[] s = Numeric.toBytesPadded(sig.s, 32);
 
         return new SignatureData(v, r, s);
-    }
-
-    private static ECDSASignature sign(byte[] transactionHash, BigInteger privateKey) {
-        ECDSASigner signer = new ECDSASigner(new HMacDSAKCalculator(new SHA256Digest()));
-
-        ECPrivateKeyParameters privKey = new ECPrivateKeyParameters(privateKey, CURVE);
-        signer.init(true, privKey);
-        BigInteger[] components = signer.generateSignature(transactionHash);
-
-        return new ECDSASignature(components[0], components[1]).toCanonicalised();
     }
 
     /**
@@ -228,48 +217,6 @@ public class Sign {
             privKey = privKey.mod(CURVE.getN());
         }
         return new FixedPointCombMultiplier().multiply(CURVE.getG(), privKey);
-    }
-
-    private static class ECDSASignature {
-        private final BigInteger r;
-        private final BigInteger s;
-
-        ECDSASignature(BigInteger r, BigInteger s) {
-            this.r = r;
-            this.s = s;
-        }
-
-        /**
-         * Returns true if the S component is "low", that means it is below
-         * {@link Sign#HALF_CURVE_ORDER}. See
-         * <a href="https://github.com/bitcoin/bips/blob/master/bip-0062.mediawiki#Low_S_values_in_signatures">
-         * BIP62</a>.
-         */
-        public boolean isCanonical() {
-            return s.compareTo(HALF_CURVE_ORDER) <= 0;
-        }
-
-        /**
-         * Will automatically adjust the S component to be less than or equal to half the curve
-         * order, if necessary. This is required because for every signature (r,s) the signature
-         * (r, -s (mod N)) is a valid signature of the same message. However, we dislike the
-         * ability to modify the bits of a Bitcoin transaction after it's been signed, as that
-         * violates various assumed invariants. Thus in future only one of those forms will be
-         * considered legal and the other will be banned.
-         */
-        public ECDSASignature toCanonicalised() {
-            if (!isCanonical()) {
-                // The order of the curve is the number of valid points that exist on that curve.
-                // If S is in the upper half of the number of valid points, then bring it back to
-                // the lower half. Otherwise, imagine that
-                //    N = 10
-                //    s = 8, so (-8 % 10 == 2) thus both (r, 8) and (r, 2) are valid solutions.
-                //    10 - 8 == 2, giving us always the latter solution, which is canonical.
-                return new ECDSASignature(r, CURVE.getN().subtract(s));
-            } else {
-                return this;
-            }
-        }
     }
 
     public static class SignatureData {
